move DeveloperQuerySortButtonComponent to Developers namespace
